+++ CHANGELOG.md
@@ -102,6 +102,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix WeightScalar - #5530 by @koradon
 - Add OrderFulfill mutation - #5525 by @fowczarek
 - Add home page - #5494 by @IKarbowiak and @dominik-zeglen
+- Extend errors in OrderFulfill mutation - #5553 by @fowczarek
 - Refactor OrderCancel mutation for many warehouses - #5554 by @fowczarek
 
 ## 2.9.0